+++ src/core/user_message_manager.hpp
@@ -18,14 +18,14 @@ public:
 	UserMessageManager() = default;
 	~UserMessageManager() = default;
 
-	bool HookUserMessage(int messageId, UserMessageCallback callback, HookMode mode);
-	bool UnhookUserMessage(int messageId, UserMessageCallback callback, HookMode mode);
+	bool HookUserMessage(uint16_t messageId, UserMessageCallback callback, HookMode mode);
+	bool UnhookUserMessage(uint16_t messageId, UserMessageCallback callback, HookMode mode);
 
 	poly::ReturnAction Hook_PostEvent(poly::Params& params, int count, poly::Return& ret, HookMode node);
 	ResultType ExecuteMessageCallbacks(INetworkMessageInternal* pEvent, CNetMessage* pData, int nClientCount, uint64* clients, HookMode mode);
 
 private:
-	std::unordered_map<int, UserMessageHook> m_hooksMap;
+	std::unordered_map<uint16_t, UserMessageHook> m_hooksMap;
 	std::array<ListenerManager<UserMessageCallback>, 2> m_globalCallbacks;
 	std::mutex m_registerCmdLock;
 };
